requirements.txt
flask>=3.0.3
flask-restx==1.3.0
flask-cors==5.0.0
gunicorn==22.0.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
pyjwt>=2.9.0
cryptography>=43.0.0
bcrypt==4.1.3
prophet==1.1.5
scikit-learn==1.4.2
scipy==1.13.0
statsmodels==0.14.2
pandas==2.2.2
numpy<2.0
flask-limiter==3.7.0
python-dotenv==1.0.1
marshmallow==3.21.2
weasyprint==62.3
firebase-admin==6.5.0
Flask-SQLAlchemy==3.1.1
cmdstanpy==1.2.3
pytesseract==0.3.10
pillow>=10.4.0
flask-socketio==5.3.6
pyotp==2.9.0
bleach==6.3.0
xgboost==2.0.3
torch>=2.3.0
torchvision>=0.18.0
ultralytics>=8.2.0
transformers>=4.40.0
shap==0.44.1
pymc==5.10.4
sentence-transformers==2.5.1
faiss-cpu==1.8.0
onnxruntime==1.17.1